+++ src/main/java/org/appjam/bongbaek/global/jwt/util/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package org.appjam.bongbaek.global.jwt;
+package org.appjam.bongbaek.global.jwt.util;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -17,13 +17,15 @@ import org.springframework.web.filter.OncePerRequestFilter;
 import java.io.IOException;
 import java.util.UUID;
 
-import static org.appjam.bongbaek.global.jwt.JwtValidationType.VALID_JWT;
+import static org.appjam.bongbaek.global.jwt.enums.JwtValidationType.VALID_JWT;
 
 @Component
 @Slf4j
 @RequiredArgsConstructor
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
     private final JwtProvider jwtProvider;
+    private final JwtValidator jwtValidator;
+    private final JwtParser jwtParser;
 
     @Override
     protected void doFilterInternal(@NonNull HttpServletRequest request,
@@ -33,8 +35,8 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
         try {
             final String token = getJwtFromRequest(request);
 
-            if (token != null && jwtProvider.validateToken(token) == VALID_JWT) {
-                UUID memberId = jwtProvider.getUserFromJwt(token);
+            if (token != null && jwtValidator.validateToken(token) == VALID_JWT) {
+                UUID memberId = jwtParser.getUserFromJwt(token);
                 MemberAuthentication authentication = MemberAuthentication.createMemberAuthentication(memberId);
                 authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
                 SecurityContextHolder.getContext().setAuthentication(authentication); // NOTE: 메타데이터, 스레드에 인증 저장
